Change int to size_t, fix ui call

+++ src/headers/ui/test_menu/teacher.h
@@ -8,7 +8,7 @@ namespace ui::test_menu
     class Teacher : public TestMenu
     {
     public:
-        void show(string test_name, string username, int amount_of_questions, unordered_map<string, int> users_points,
+        void show(string test_name, string username, size_t amount_of_questions, unordered_map<string, int> users_points,
                   string selected_answer) override;
     };
 }
