+++ src/ui/test_menu/student.cpp
@@ -1,6 +1,6 @@
 #include "../../headers/ui/test_menu/student.h"
 
-void ui::test_menu::Student::show(string test_name, string username, int amount_of_questions, unordered_map<string, int> users_points,
+void ui::test_menu::Student::show(string test_name, string username, size_t amount_of_questions, unordered_map<string, int> users_points,
                                   string selected_answer)
 {
     // TODO: Show student test menu.
